HW2/requirements.txt
absl-py==2.1.0
astroid==3.3.9
black==24.10.0
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1739515848642/work/certifi
click==8.1.8
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
contourpy @ file:///home/conda/feedstock_root/build_artifacts/contourpy_1727293517607/work
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1733332471406/work
Cython @ file:///home/conda/feedstock_root/build_artifacts/cython_1739227939853/work
dill==0.3.9
filelock==3.17.0
fonttools @ file:///home/conda/feedstock_root/build_artifacts/fonttools_1738940303262/work
fsspec==2025.3.0
grpcio==1.70.0
importlib_metadata==8.6.1
importlib_resources @ file:///home/conda/feedstock_root/build_artifacts/importlib_resources_1736252299705/work
isort==6.0.1
Jinja2==3.1.6
kiwisolver @ file:///home/conda/feedstock_root/build_artifacts/kiwisolver_1725459266648/work
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.4
mccabe==0.7.0
mpmath==1.3.0
munkres==1.1.4
mypy-extensions==1.0.0
networkx==3.2.1
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1732314280888/work/dist/numpy-2.0.2-cp39-cp39-linux_x86_64.whl#sha256=62d98eb3da9f13e6b227c430d01026b7427f341b3fdcb838430f2a9e520417b1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.11.0
opencv-python-headless==4.11.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1733203243479/work
pandas @ file:///home/conda/feedstock_root/build_artifacts/pandas_1736810577256/work
pathspec==0.12.1
pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1735929703139/work
platformdirs==4.3.7
protobuf==6.30.0
pycocotools @ file:///home/conda/feedstock_root/build_artifacts/pycocotools_1731600557898/work
pyink==24.10.1
PyLink==0.3.3
pylint==3.3.6
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1735698276679/work
pyserial==3.5
PySide6==6.8.2
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1733215673016/work
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1706886791323/work
shiboken6==6.8.2
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tomli==2.2.1
tomlkit==0.13.2
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1732615921868/work
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1735661334605/work
triton==3.2.0
typing_extensions==4.12.2
tzdata @ file:///home/conda/feedstock_root/build_artifacts/python-tzdata_1737541069190/work
unicodedata2 @ file:///home/conda/feedstock_root/build_artifacts/unicodedata2_1736692503055/work
Werkzeug==3.1.3
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1732827521216/work
